Added condition when closing Redis server

+++ desktop/electron-app/src/main.js
@@ -75,7 +75,9 @@ app.on('window-all-closed', () => {
 // Quit the app when closed
 app.on('will-quit', () => {
   server.close();
-  redisServer.kill();
+  if (redisServer) {
+    redisServer.kill();
+  }
 });
 
 function startRedisServer() {
